add scene with spheres

+++ include/camera.h
@@ -3,8 +3,8 @@
 
 class Camera {
     public:
-        Camera(std::array<int, 3> position, int vw, int vh, int d);
-        std::array<int, 3> _position;
+        Camera(std::array<double, 3> position, int vw, int vh, int d);
+        std::array<double, 3> _position;
         //Viewport size
         int _vw;
         int _vh;
